requirements.txt
Django>=4.1,<4.2
django-environ
requests==2.28.2
django_extensions==3.2.1
django-jsonfield==1.4.1
mysqlclient==2.1.1
environ==1.0
whitenoise
boto3
uwsgi
django-storages
redis
djangorestframework
markdown
django-filter
drf-yasg